Made Layout Content Client Side as required.
